Corrected build issue when including as subproject
Some paths related to the documents directory in install
statements were relative to CMAKE_SOURCE_DIR which changes
when including the materialx repo as a sub module in another
cmake file causing the install to fail.
This change moves these path to be relative to
CMAKE_CURRENT_SOURCE_DIR which should be stable.